backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
requests
cryptography
pypdf
python-docx
